restructure files + reading order logic with columns
